requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cssselect==1.1.0
cssutils==2.3.0
idna==2.10
importlib-metadata==4.0.1
lxml==4.6.3
premailer==3.8.0
python-dotenv==0.17.1
requests==2.25.1
typing-extensions==3.10.0.0
urllib3==1.26.4
yagmail==0.14.256
zipp==3.4.1
